mentally-backend/requirements.txt
uvicorn
fastapi
python-dotenv
pydantic
firebase
firebase_admin
pyrebase4
requests